bigint

+++ src/main/java/org/hhhonzik/starrocks/TrinoHLLCardinality.java
@@ -1,5 +1,6 @@
 package org.hhhonzik.starrocks;
 
+import java.math.BigInteger;
 import java.nio.Buffer;
 
 // import com.facebook.airlift.stats.cardinality.HyperLogLog;
@@ -7,8 +8,8 @@ import java.nio.Buffer;
 
 public class TrinoHLLCardinality {
 
-        public final Long evaluate(Buffer obj) {
-        return 123L;
+        public final BigInteger evaluate(Buffer obj) {
+        return BigInteger.valueOf(123);
     }
 }
